lint系パッケージのアップデート

+++ src/lib/firestore.ts
@@ -6,7 +6,7 @@ import {
   DeleteDiaryAction,
   Diary,
   GetDiariesAction,
-  GetDiaryAction
+  GetDiaryAction,
 } from "../store/diary/types";
 import { User } from "../store/user/types";
 
@@ -14,7 +14,7 @@ export const fs = firebase.firestore();
 
 export async function getUserFromFirestore({
   firestore,
-  userId
+  userId,
 }: {
   firestore: FirebaseFirestore.Firestore;
   userId: string;
@@ -23,17 +23,17 @@ export async function getUserFromFirestore({
     .collection(`users`)
     .doc(userId)
     .get()
-    .then(doc => doc.data() as User);
+    .then((doc) => doc.data() as User);
   return {
     uid: userId,
     name: user.name || "No Name",
-    picture: user.picture
+    picture: user.picture,
   };
 }
 
 async function setDiaryUserToFireStore({
   firestore = fs,
-  user
+  user,
 }: {
   firestore?: FirebaseFirestore.Firestore | firebase.firestore.Firestore;
   user: User;
@@ -47,34 +47,31 @@ async function setDiaryUserToFireStore({
 export async function createDiaryToFirestore({
   firestore = fs,
   user,
-  diary
+  diary,
 }: {
   firestore?: FirebaseFirestore.Firestore | firebase.firestore.Firestore;
   user: User;
   diary: Diary;
-}) {
+}): Promise<void> {
   setDiaryUserToFireStore({ firestore, user });
-  firestore
-    .collection(`/users/${user.uid}/diaries`)
-    .doc(`${diary.id}`)
-    .set({
-      id: diary.id,
-      title: diary.title,
-      body: diary.body,
-      lastEdited: new Date()
-    });
+  firestore.collection(`/users/${user.uid}/diaries`).doc(`${diary.id}`).set({
+    id: diary.id,
+    title: diary.title,
+    body: diary.body,
+    lastEdited: new Date(),
+  });
 }
 
 export async function getDiaryFromFirestore({
   firestore,
   userId,
-  diaryId
+  diaryId,
 }: GetDiaryAction["payload"]): Promise<Diary | undefined> {
   const diaryData = await firestore
     .collection(`users/${userId}/diaries/`)
     .doc(`${diaryId}`)
     .get()
-    .then(doc => doc.data());
+    .then((doc) => doc.data());
   if (!diaryData) return undefined;
   return {
     id: diaryData.id,
@@ -83,20 +80,20 @@ export async function getDiaryFromFirestore({
     lastEdited: utcToZonedTime(
       fromUnixTime(diaryData.lastEdited.seconds),
       "Asia/Tokyo"
-    ).toISOString()
+    ).toISOString(),
   };
 }
 
 export async function getDiariesFromFirestore({
   firestore,
-  userId
-}: GetDiariesAction["payload"]) {
-  const diariesData: any[] = [];
+  userId,
+}: GetDiariesAction["payload"]): Promise<Diary[]> {
+  const diariesData: Diary[] = [];
   await firestore
     .collection(`users/${userId}/diaries`)
     .get()
-    .then((collections: any) => {
-      collections.forEach((doc: any) => {
+    .then((collections) => {
+      collections.forEach((doc) => {
         const data = doc.data();
         diariesData.push({
           id: data.id,
@@ -106,7 +103,7 @@ export async function getDiariesFromFirestore({
           lastEdited: utcToZonedTime(
             fromUnixTime(data.lastEdited.seconds),
             "Asia/Tokyo"
-          ).toISOString()
+          ).toISOString(),
         });
       });
     });
@@ -116,10 +113,7 @@ export async function getDiariesFromFirestore({
 export async function deleteDiaryFromFirestore({
   firestore,
   userId,
-  diaryId
-}: DeleteDiaryAction["payload"]) {
-  await firestore
-    .collection(`users/${userId}/diaries/`)
-    .doc(diaryId)
-    .delete();
+  diaryId,
+}: DeleteDiaryAction["payload"]): Promise<void> {
+  await firestore.collection(`users/${userId}/diaries/`).doc(diaryId).delete();
 }
